+++ Stripe/STPAddressViewModel.m
@@ -17,7 +17,7 @@
 @interface STPAddressViewModel()<STPAddressFieldTableViewCellDelegate>
 @property (nonatomic) BOOL isBillingAddress;
 @property (nonatomic) STPBillingAddressFields requiredBillingAddressFields;
-@property (nonatomic) PKAddressField requiredShippingAddressFields;
+@property (nonatomic) NSSet<STPContactField> *requiredShippingAddressFields;
 @property (nonatomic) NSArray<STPAddressFieldTableViewCell *> *addressCells;
 @property (nonatomic) BOOL showingPostalCodeCell;
 @property (nonatomic) BOOL geocodeInProgress;
@@ -58,19 +58,19 @@
     return self;
 }
 
-- (instancetype)initWithRequiredShippingFields:(PKAddressField)requiredShippingAddressFields {
+- (instancetype)initWithRequiredShippingFields:(NSSet<STPContactField> *)requiredShippingAddressFields {
     self = [super init];
     if (self) {
         _isBillingAddress = NO;
         _requiredShippingAddressFields = requiredShippingAddressFields;
         NSMutableArray *cells = [NSMutableArray new];
-        if (requiredShippingAddressFields & PKAddressFieldName) {
+        if ([requiredShippingAddressFields containsObject:STPContactFieldName]) {
             [cells addObject:[[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypeName contents:@"" lastInList:NO delegate:self]];
         }
-        if (requiredShippingAddressFields & PKAddressFieldEmail) {
+        if ([requiredShippingAddressFields containsObject:STPContactFieldEmailAddress]) {
             [cells addObject:[[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypeEmail contents:@"" lastInList:NO delegate:self]];
         }
-        if (requiredShippingAddressFields & PKAddressFieldPostalAddress) {
+        if ([requiredShippingAddressFields containsObject:STPContactFieldPostalAddress]) {
             NSMutableArray *postalCells = [@[
                                              [[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypeName contents:@"" lastInList:NO delegate:self],
                                              [[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypeLine1 contents:@"" lastInList:NO delegate:self],
@@ -80,12 +80,12 @@
                                              [[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypeState contents:@"" lastInList:NO delegate:self],
                                              [[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypeCountry contents:_addressFieldTableViewCountryCode lastInList:NO delegate:self],
                                              ] mutableCopy];
-            if (requiredShippingAddressFields & PKAddressFieldName) {
+            if ([requiredShippingAddressFields containsObject:STPContactFieldName]) {
                 [postalCells removeObjectAtIndex:0];
             }
             [cells addObjectsFromArray:postalCells];
         }
-        if (requiredShippingAddressFields & PKAddressFieldPhone) {
+        if ([requiredShippingAddressFields containsObject:STPContactFieldPhoneNumber]) {
             [cells addObject:[[STPAddressFieldTableViewCell alloc] initWithType:STPAddressFieldTypePhone contents:@"" lastInList:NO delegate:self]];
         }
         STPAddressFieldTableViewCell *lastCell = [cells lastObject];
@@ -159,7 +159,7 @@
         return self.requiredBillingAddressFields == STPBillingAddressFieldsFull;
     }
     else {
-        return (self.requiredShippingAddressFields & PKAddressFieldPostalAddress) == PKAddressFieldPostalAddress;
+        return [self.requiredShippingAddressFields containsObject:STPContactFieldPostalAddress];
     }
 }